Add files via upload
I made the nurseView as well as a couple methods such as displayRecordVitalScene so it can update the patient information file. It was added on top of the previous code so I will test after class with the rest of the other codes. Feel free to read through comments and tell me anything thats missing I'll be ready to fix.