Add files via upload
Okkkk I think this is a good changeup from the previous code. It is certainly running better and is uploading the vitals onto the same files created. Feel free to test it out, I just need to work on the view patient history, add back buttons to the record vitals page, and some other necessary tweaks.